Fix import after tiktorch refactor
