+++ ilastik/workflows/neuralNetwork/_localLauncher.py
@@ -30,7 +30,7 @@ from typing import List, Optional
 
 import grpc
 import psutil
-from tiktorch.proto.inference_pb2 import Empty
+from tiktorch.proto.utils_pb2 import Empty
 from tiktorch.proto.inference_pb2_grpc import FlightControlStub
 
 logger = logging.getLogger(__name__)
